// src/server/enum/weekday.enum.ts
export enum EnumWeekday {
    ENG_MONDAY = "monday",
    ENG_TUESDAY = "tuesday",
    ENG_WEDNESDAY = "wednesday",
    ENG_THURSDAY = "thursday",
    ENG_FRIDAY = "friday",
    ENG_SATURDAY = "saturday",
    ENG_SUNDAY = "sunday",

    SE_MANDAG = "måndag",
    SE_TISDAG = "tisdag",
    SE_ONSDAG = "onsdag",
    SE_TORSDAG = "torsdag",
    SE_FREDAG = "fredag",
    SE_LORDAG = "lördag",
    SE_SONDAG= "söndag",
}